refactor(ListSeparator): avoid circular dependency
