+++ src/interface.ts
@@ -16,10 +16,7 @@ export * from './components/help/help.types';
 export * from './components/icon/icon.types';
 export * from './components/info-tile/info-tile.types';
 export * from './components/input-field/input-field.types';
-export {
-    ListComponent,
-    ListSeparator,
-} from './components/list/list-item.types';
+export { ListComponent } from './components/list/list-item.types';
 export * from './components/list/list.types';
 export * from './components/menu/menu.types';
 export * from './components/menu-list/menu-list.types';
@@ -32,3 +29,4 @@ export * from './components/tab-panel/tab-panel.types';
 export * from './components/table/table.types';
 export * from './global/shared-types/link.types';
 export * from './global/shared-types/icon.types';
+export * from './global/shared-types/separator.types';
